Seperate components in Progress and help distinguish tapping and panning events in customed calendars
