+++ components/main/screens/Progress/Progress.Presentational.js
@@ -6,12 +6,13 @@ import {
   TouchableOpacity,
   Modal,
   Dimensions,
-  Animated
+  Animated,
+  StyleSheet,
 } from 'react-native';
 import { PanGestureHandler, State } from 'react-native-gesture-handler'
 import { Map } from 'immutable'
 import { StackedBarChart, XAxis, YAxis } from 'react-native-svg-charts'
-
+import PointEarnedSection from './components/point-earned-section/PointEarnedSection'
 export default class Progress extends React.PureComponent {
   year_array = [new Date().getFullYear(), new Date().getFullYear() + 1, new Date().getFullYear() - 1]
 
@@ -54,25 +55,7 @@ export default class Progress extends React.PureComponent {
     return (
       <>
         <ScrollView>
-          <Text
-            style={{
-              width: 102,
-              height: 20,
-              marginTop: 16,
-              marginLeft: 16,
-              fontStyle: "normal",
-              fontWeight: "normal",
-              fontSize: 16,
-              lineHeight: 20,
-            }}
-          >
-            Point earned:
-          </Text>
-
-          <Calendar
-            month={this.state.month}
-            year={this.state.year}
-            toggleChooseMonth={this.toggleChooseMonth}
+          <PointEarnedSection
             {...this.props}
           />
 
@@ -84,924 +67,14 @@ export default class Progress extends React.PureComponent {
             year={this.state.year}
           />
 
-          <Charts />
+          <ChartSection />
         </ScrollView >
-
-        {this.state.choose_month_bool ?
-          <Modal
-            transparent={true}
-          >
-            <View
-              style={{
-                flex: 1,
-                position: "relative",
-                justifyContent: "center",
-                alignItems: "center"
-              }}
-            >
-              <TouchableOpacity
-                style={{
-                  flex: 1,
-                  width: Dimensions.get("window").width,
-                  backgroundColor: "black",
-                  opacity: 0.5
-                }}
-
-                onPress={this.dismissChooseMonth}
-              >
-
-              </TouchableOpacity>
-
-              <MonthCalendar
-                chooseMonth={this.chooseMonth}
-                dismissChooseMonth={this.dismissChooseMonth}
-                year_array={this.year_array}
-              />
-            </View>
-          </Modal>
-          :
-          null
-        }
-
       </>
     );
   }
 }
 
-class Calendar extends React.PureComponent {
-  state = {
-    month_data_array: []
-  }
 
-  getWeek = (date) => {
-    let target = new Date(date);
-    let dayNr = (date.getDay() + 6) % 7;
-    target.setDate(target.getDate() - dayNr + 3);
-    let firstThursday = target.valueOf();
-    target.setMonth(0, 1);
-    if (target.getDay() != 4) {
-      target.setMonth(0, 1 + ((4 - target.getDay()) + 7) % 7);
-    }
-    return 1 + Math.ceil((firstThursday - target) / 604800000);
-  }
-
-  getMonday = (date) => {
-    let dayInWeek = new Date(date).getDay()
-    let diff = dayInWeek === 0 ? 6 : dayInWeek - 1
-    return new Date(new Date(date).getTime() - (diff * 86400 * 1000))
-  }
-
-  getNoWeekInMonth = (date) => {
-    let nearest_monday = this.getMonday(date)
-    let first_moday_of_month = this.getMonday(new Date(date.getFullYear(), date.getMonth(), 7))
-
-    return Math.floor((nearest_monday - first_moday_of_month) / 7) + 1
-  }
-
-  initDaysInMonth = (month, year) => {
-    let month_data_array = []
-
-    let first_monday = this.getMonday(new Date(year, month, 1)),
-      last_monday = this.getMonday(new Date(year, month + 1, 0)),
-
-      first_week = this.getWeek(first_monday),
-      last_week = this.getWeek(last_monday),
-      monday_of_week = first_monday
-
-    if (first_week > last_week) {
-      last_week += first_week
-
-      for (let i = first_week; i <= last_week; i++) {
-        let week = i % first_week === 0 ? first_week : i % first_week
-        month_data_array.push({
-          week_row: true,
-          week,
-          start_day: monday_of_week.getDate(),
-          start_month: monday_of_week.getMonth(),
-          start_year: monday_of_week.getFullYear(),
-        })
-
-        for (let j = 0; j < 7; j++) {
-          let date = new Date(monday_of_week.getTime() + 86400 * 1000 * j)
-          month_data_array.push({
-            day: date.getDate(),
-            month: date.getMonth(),
-            year: date.getFullYear(),
-            day_in_week: date.getDay(),
-            week
-          })
-        }
-
-        monday_of_week = new Date(monday_of_week.getTime() + 86400 * 1000 * 7)
-      }
-    }
-
-    else {
-      for (let i = first_week; i <= last_week; i++) {
-        month_data_array.push({
-          week_row: true,
-          week: i,
-          start_day: monday_of_week.getDate(),
-          start_month: monday_of_week.getMonth(),
-          start_year: monday_of_week.getFullYear(),
-        })
-
-        for (let j = 0; j < 7; j++) {
-          let date = new Date(monday_of_week.getTime() + 86400 * 1000 * j)
-          month_data_array.push({
-            day: date.getDate(),
-            month: date.getMonth(),
-            year: date.getFullYear(),
-            day_in_week: date.getDay(),
-            week: i,
-          })
-        }
-
-        monday_of_week = new Date(monday_of_week.getTime() + 86400 * 1000 * 7)
-
-      }
-    }
-
-    this.setState({
-      month_data_array: [...month_data_array]
-    })
-  }
-
-  componentDidMount() {
-    this.initDaysInMonth(this.props.month, this.props.year)
-  }
-
-  componentDidUpdate(prevProps, prevState) {
-    if (this.props.month !== prevProps.month || this.props.year !== prevProps.year) {
-      this.initDaysInMonth(this.props.month, this.props.year)
-    }
-  }
-
-  render() {
-    return (
-      <View
-        style={{
-          width: 400,
-        }}
-      >
-        <MonthHolder
-          month_data_array={this.state.month_data_array}
-          {...this.props}
-        />
-      </View>
-    )
-  }
-}
-
-class MonthHolder extends React.Component {
-
-  day_texts = ["M", "T", "W", "T", "F", "S", "S"]
-  month_texts = ["January", "Febuary", "March", "April", "May", "June", "July", "August", "September", "October", "November", "December"]
-
-  state = {
-    day_text_array: null,
-    month_points: 0,
-  }
-
-  _toggleChooseMonth = () => {
-    this.props.toggleChooseMonth()
-  }
-
-  updateMonthPoints = () => {
-    let date = new Date(this.props.year, this.props.month).getTime(),
-      month_stats = Map(this.props.month_stats)
-    if (month_stats.has(date)) {
-      let stat = month_stats.get(date),
-        total_points = stat.current.reduce(((total, amount) => total + amount), 0)
-
-      this.setState({
-        month_points: total_points
-      })
-    }
-
-    else {
-      this.setState({
-        month_points: 0
-      })
-    }
-  }
-
-  componentDidMount() {
-    this.updateMonthPoints()
-  }
-
-  componentDidUpdate(prevProps, prevState) {
-    if (this.props.month_data_array !== prevProps.month_data_array) {
-      let day_text_array = []
-
-      day_text_array.push(
-        <View
-          key={`week-text`}
-          style={{
-            height: 40,
-            width: 400 / 8,
-            justifyContent: "center",
-            alignItems: "center",
-          }}
-        >
-          <Text>
-            Week
-          </Text>
-        </View>
-      )
-
-      for (let i = 0; i < 7; i++) {
-        day_text_array.push(
-          <View
-            key={`day-text-${i}`}
-            style={{
-              height: 40,
-              width: 400 / 8,
-              justifyContent: "center",
-              alignItems: "center",
-            }}
-          >
-            <Text>
-              {this.day_texts[i]}
-            </Text>
-          </View>
-        )
-      }
-
-      this.setState({
-        day_text_array: [...day_text_array],
-      })
-    }
-
-    if (this.props.month_stats !== prevProps.month_stats || (this.props.month !== prevProps.month) || (this.props.year !== prevProps.year)) {
-      this.updateMonthPoints()
-    }
-
-  }
-
-
-  render() {
-    return (
-      <View
-        style={{
-          flex: 1,
-          width: 400,
-        }}
-      >
-        <View
-          style={{
-            height: 50,
-            justifyContent: "center",
-            alignItems: "center",
-          }}
-        >
-          <TouchableOpacity
-            style={{
-              flexDirection: "row",
-            }}
-            onPress={this._toggleChooseMonth}
-          >
-            <Text>
-              {`${this.month_texts[this.props.month]} - ${this.props.year}`}
-            </Text>
-
-            {this.state.month_points > 0 ?
-              <Text
-                style={{
-                  marginLeft: 5,
-                }}
-              >
-                {`${this.state.month_points}pt`}
-              </Text>
-              :
-              null
-            }
-
-          </TouchableOpacity>
-        </View>
-
-        <View
-          style={{
-            flexDirection: "row",
-            flexWrap: "wrap",
-          }}
-        >
-          {this.state.day_text_array}
-          {this.props.month_data_array.map((data, index) => {
-            if (data.week_row) {
-              return (
-                <Week
-                  key={`week-${index}`}
-                  data={data}
-                  {... this.props}
-                />
-              )
-            }
-            return (
-              <Day
-                key={`day-${index}`}
-                data={data}
-                {... this.props}
-              />
-            )
-          })
-          }
-        </View>
-      </View>
-    )
-  }
-}
-
-class Day extends React.PureComponent {
-
-  state = {
-    day_points: 0,
-  }
-
-  updateDayPoints = () => {
-    let { data } = this.props,
-      date = new Date(data.year, data.month, data.day).getTime(),
-      day_stats = Map(this.props.day_stats)
-
-    if (day_stats.has(date)) {
-      let stat = day_stats.get(date),
-        total_points = stat.current.reduce(((total, amount) => total + amount), 0)
-
-      this.setState({
-        day_points: total_points
-      })
-    }
-
-    else {
-      this.setState({
-        day_points: 0
-      })
-    }
-  }
-
-  componentDidMount() {
-    this.updateDayPoints()
-  }
-
-  componentDidUpdate(prevProps, prevState) {
-    if (this.props.day_stats !== prevProps.day_stats || this.props.data !== prevProps.data) {
-      this.updateDayPoints()
-    }
-
-  }
-
-  render() {
-    return (
-      <TouchableOpacity
-        style={{
-          height: 100,
-          width: 400 / 8,
-          alignItems: "center"
-        }}
-      // onPress={this._onPress}
-      >
-        <View
-          style={{
-            width: 32,
-            height: 77,
-            borderRadius: 17,
-            backgroundColor: this.state.day_points > 0 ? "#C4C4C4" : "white",
-            paddingHorizontal: 3,
-            paddingVertical: 3,
-            alignItems: "center",
-          }}
-        >
-          <View
-            style={{
-              width: 26,
-              height: 26,
-              borderRadius: 13,
-              justifyContent: "center",
-              alignItems: "center",
-              backgroundColor: "white"
-            }}
-          >
-            <Text>
-              {this.props.data.day}
-            </Text>
-          </View>
-
-          {this.state.day_points > 0 ?
-            <Text
-              style={{
-                marginTop: 20,
-              }}
-            >
-              {`${this.state.day_points}pt`}
-            </Text>
-            :
-
-            null
-          }
-
-        </View>
-      </TouchableOpacity>
-    )
-  }
-}
-
-class Week extends React.PureComponent {
-  state = {
-    week_points: 0,
-  }
-
-  updateWeekPoints = () => {
-    let { data } = this.props,
-      date = new Date(data.start_year, data.start_month, data.start_day).getTime(),
-      week_stats = Map(this.props.week_stats)
-    if (week_stats.has(date)) {
-
-      let stat = week_stats.get(date),
-        total_points = stat.current.reduce(((total, amount) => total + amount), 0)
-
-      this.setState({
-        week_points: total_points
-      })
-    }
-
-    else {
-      this.setState({
-        week_points: 0
-      })
-    }
-  }
-
-  componentDidMount() {
-    this.updateWeekPoints()
-  }
-
-  componentDidUpdate(prevProps, prevState) {
-    if (this.props.week_stats !== prevProps.week_stats || this.props.data !== prevProps.data) {
-      this.updateWeekPoints()
-    }
-  }
-
-  render() {
-    return (
-      <TouchableOpacity
-        style={{
-          height: 100,
-          width: 400 / 8,
-          alignItems: "center"
-        }}
-      // onPress={this._onPress}
-      >
-        <View
-          style={{
-            width: 32,
-            height: 77,
-            borderRadius: 17,
-            backgroundColor: this.state.week_points > 0 ? "#54BAAC" : "white",
-            paddingHorizontal: 3,
-            paddingVertical: 3,
-            alignItems: "center",
-          }}
-        >
-          <View
-            style={{
-              width: 26,
-              height: 26,
-              borderRadius: 13,
-              justifyContent: "center",
-              alignItems: "center",
-              backgroundColor: "white"
-            }}
-          >
-            <Text>
-              {this.props.data.week}
-            </Text>
-          </View>
-
-          {this.state.week_points > 0 ?
-            <Text
-              style={{
-                marginTop: 20,
-              }}
-            >
-              {`${this.state.week_points}pt`}
-            </Text>
-            :
-            null
-          }
-
-        </View>
-      </TouchableOpacity>
-    )
-  }
-}
-
-
-class MonthCalendar extends React.PureComponent {
-  dimension_width = 350
-
-  old_translateX = 0
-
-  main_index = 0
-
-  translateX_array = [new Animated.Value(0), new Animated.Value(this.dimension_width), new Animated.Value(-this.dimension_width)]
-
-  record_translateX_array = [0, this.dimension_width, -this.dimension_width]
-
-  opacity_array = [new Animated.Value(1), new Animated.Value(1), new Animated.Value(1)]
-
-  state = {
-    year_array: [0, 0, 0]
-  }
-
-  handleMonthYearWhenSwipe = (swipe_direction, main_index) => {
-    let year_array = [... this.state.year_array]
-
-    if (swipe_direction === 1) {
-      if (main_index === 0) {
-        year_array[2] = year_array[main_index] - 1
-      }
-
-      else {
-        year_array[(main_index - 1) % 3] = year_array[main_index] - 1
-      }
-    }
-
-    else {
-      year_array[(main_index + 1) % 3] = year_array[main_index] + 1
-    }
-
-    this.setState({
-      year_array: [...year_array]
-    })
-
-  }
-
-
-  _onGestureEvent = Animated.event(
-    [
-      {},
-    ],
-    {
-      listener: ({ nativeEvent }) => {
-        this.record_translateX_array.forEach((translate, index, arr) => {
-          arr[index] += nativeEvent.translationX - this.old_translateX
-        })
-
-        this.translateX_array.forEach((translate, index, arr) => {
-          arr[index].setValue(this.record_translateX_array[index])
-        })
-
-        this.old_translateX = nativeEvent.translationX
-
-        if (this.record_translateX_array[this.main_index] >= -120 && this.record_translateX_array[this.main_index] <= 120) {
-          this.opacity_array[(this.main_index + 1) % 3].setValue(1)
-
-          if (this.main_index === 0) {
-            this.opacity_array[2].setValue(1)
-          }
-          else {
-            this.opacity_array[(this.main_index - 1) % 3].setValue(1)
-          }
-        }
-
-        this.opacity_array[this.main_index].setValue((this.dimension_width - Math.abs(this.record_translateX_array[this.main_index])) / (this.dimension_width * 1.3))
-      }
-    }
-  )
-
-  handleAnimation = (main_index) => {
-    if (this.record_translateX_array[main_index] >= -120 && this.record_translateX_array[main_index] <= 120) {
-      this.opacity_array[main_index].setValue(1)
-
-      this.record_translateX_array[main_index] = 0
-      this.record_translateX_array[(main_index + 1) % 3] = this.dimension_width
-
-      if (main_index - 1 < 0) {
-        main_index = 6
-      }
-
-      this.record_translateX_array[(main_index - 1) % 3] = - this.dimension_width
-
-      let animation_array = this.translateX_array.map((translate, index) =>
-        Animated.spring(this.translateX_array[index], {
-          toValue: this.record_translateX_array[index]
-        })
-      )
-
-      Animated.parallel(
-        animation_array,
-        {
-          stopTogether: true
-        }
-      ).start()
-    }
-
-    else if (this.record_translateX_array[main_index] < -120) {
-
-      this.record_translateX_array[main_index] = - this.dimension_width
-      this.record_translateX_array[(main_index + 1) % 3] = 0
-
-      if (main_index - 1 < 0) {
-        main_index = 6
-      }
-
-      this.record_translateX_array[(main_index - 1) % 3] = this.dimension_width
-
-      this.translateX_array[(main_index - 1) % 3].setValue(this.dimension_width)
-
-      let animation_array = this.translateX_array.map((translate, index) => {
-        if (index !== (main_index - 1) % 3)
-          return (
-            Animated.spring(this.translateX_array[index], {
-              toValue: this.record_translateX_array[index]
-            })
-          )
-      })
-
-      Animated.parallel(
-        animation_array,
-        {
-          stopTogether: true
-        }
-      ).start()
-
-      this.handleMonthYearWhenSwipe(-1, this.main_index)
-
-      this.main_index = (main_index + 1) % 3
-    }
-
-    else if (this.record_translateX_array[main_index] > 120) {
-
-      this.record_translateX_array[main_index] = this.dimension_width
-
-      this.record_translateX_array[(main_index + 1) % 3] = - this.dimension_width
-
-      this.translateX_array[(main_index + 1) % 3].setValue(-this.dimension_width)
-
-      if (main_index - 1 < 0) {
-        main_index = 6
-      }
-      this.record_translateX_array[(main_index - 1) % 3] = 0
-
-
-      let animation_array = this.translateX_array.map((translate, index) => {
-        if (index !== (main_index + 1) % 3)
-          return (
-            Animated.spring(this.translateX_array[index], {
-              toValue: this.record_translateX_array[index]
-            })
-          )
-      })
-
-      Animated.parallel(
-        animation_array,
-        {
-          stopTogether: true
-        }
-      ).start()
-
-      this.handleMonthYearWhenSwipe(1, this.main_index)
-
-      this.main_index = (main_index - 1) % 3
-    }
-  }
-
-  _onHandlerStateChange = ({ nativeEvent }) => {
-    if (nativeEvent.state === State.END) {
-      this.old_translateX = 0
-
-      this.handleAnimation(this.main_index)
-    }
-  }
-
-  componentDidMount() {
-    this.setState({
-      year_array: [...this.props.year_array]
-    })
-  }
-
-  componentDidUpdate(prevProps, prevState) {
-    if (this.props.year_array !== prevProps.year_array) {
-      this.setState({
-        year_array: [...this.props.year_array]
-      })
-    }
-  }
-
-  render() {
-    return (
-      <View
-        style={{
-          width: 350,
-          height: 400,
-          borderRadius: 20,
-          backgroundColor: "white",
-          position: "absolute",
-          flexDirection: "row",
-          justifyContent: "center",
-          alignItems: "center",
-          overflow: "hidden",
-        }}
-      >
-        <PanGestureHandler
-          onGestureEvent={this._onGestureEvent}
-          onHandlerStateChange={this._onHandlerStateChange}
-        >
-          <Animated.View
-            style={{
-              flex: 1,
-              width: 350,
-              justifyContent: "center",
-              alignItems: "center",
-              transform: [{ translateX: this.translateX_array[0] }],
-              opacity: this.opacity_array[0],
-            }}
-          >
-            <MonthsInYear
-              year={this.state.year_array[0]}
-              {...this.props}
-            />
-          </Animated.View>
-        </PanGestureHandler>
-
-        <PanGestureHandler
-          onGestureEvent={this._onGestureEvent}
-          onHandlerStateChange={this._onHandlerStateChange}
-        >
-          <Animated.View
-            style={{
-              height: 400,
-              width: 350,
-              justifyContent: "center",
-              alignItems: "center",
-              transform: [{ translateX: this.translateX_array[1] }],
-              position: "absolute",
-              opacity: this.opacity_array[1],
-            }}
-          >
-            <MonthsInYear
-              year={this.state.year_array[1]}
-              {...this.props}
-            />
-          </Animated.View>
-        </PanGestureHandler>
-
-        <PanGestureHandler
-          onGestureEvent={this._onGestureEvent}
-          onHandlerStateChange={this._onHandlerStateChange}
-        >
-          <Animated.View
-            style={{
-              height: 400,
-              width: 350,
-              justifyContent: "center",
-              alignItems: "center",
-              transform: [{ translateX: this.translateX_array[2] }],
-              position: "absolute",
-              opacity: this.opacity_array[2],
-            }}
-          >
-            <MonthsInYear
-              year={this.state.year_array[2]}
-              {...this.props}
-            />
-          </Animated.View>
-        </PanGestureHandler>
-      </View>
-    )
-  }
-}
-
-class MonthsInYear extends React.PureComponent {
-
-  state = {
-    year_data_array: []
-  }
-
-  getWeek = (date) => {
-    let target = new Date(date);
-    let dayNr = (date.getDay() + 6) % 7;
-    target.setDate(target.getDate() - dayNr + 3);
-    let firstThursday = target.valueOf();
-    target.setMonth(0, 1);
-    if (target.getDay() != 4) {
-      target.setMonth(0, 1 + ((4 - target.getDay()) + 7) % 7);
-    }
-    return 1 + Math.ceil((firstThursday - target) / 604800000);
-  }
-
-  getMonday = (date) => {
-    let dayInWeek = new Date(date).getDay()
-    let diff = dayInWeek === 0 ? 6 : dayInWeek - 1
-    return new Date(new Date(date).getTime() - (diff * 86400 * 1000))
-  }
-
-  getNoWeekInMonth = (date) => {
-    let nearest_monday = this.getMonday(date)
-    let first_moday_of_month = this.getMonday(new Date(date.getFullYear(), date.getMonth(), 7))
-
-    return Math.floor((nearest_monday - first_moday_of_month) / 7) + 1
-  }
-
-  initMonthsInYear = (year) => {
-    let year_data_array = []
-    for (let i = 0; i < 12; i++) {
-      let start_day = new Date(year, i, 1),
-        last_day = new Date(year, i + 1, 0)
-
-      year_data_array.push({
-        month: i,
-        year,
-        start_week: this.getWeek(start_day),
-        end_week: this.getWeek(last_day)
-      })
-    }
-
-    this.setState({
-      year_data_array: [...year_data_array]
-    })
-  }
-
-  componentDidMount() {
-    this.initMonthsInYear(this.props.year)
-  }
-
-  componentDidUpdate(prevProps, prevState) {
-    if (this.props.year !== prevProps.year) {
-      this.initMonthsInYear(this.props.year)
-    }
-
-    if (this.props.year_array !== prevProps.year_array) {
-      console.log(this.props.year_array)
-    }
-  }
-
-  render() {
-    return (
-      <View
-        style={{
-          alignItems: "center",
-        }}
-      >
-        <Text
-          style={{
-            marginTop: 20,
-          }}
-        >
-          {this.props.year}
-        </Text>
-
-        <View
-          style={{
-            flex: 1,
-            flexDirection: "row",
-            flexWrap: "wrap",
-            marginTop: 20,
-          }}
-        >
-          {this.state.year_data_array.map((data, index) => (
-            <Month
-              data={data}
-              key={`month-in-year-${index}`}
-              {...this.props}
-            />
-          ))}
-        </View>
-      </View>
-    )
-  }
-}
-
-class Month extends React.PureComponent {
-  short_month_texts = ["Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep", "Oct", "Nov", "Dec"]
-
-  _onPress = () => {
-    this.props.chooseMonth(this.props.data.month, this.props.data.year)
-    this.props.dismissChooseMonth()
-  }
-
-  render() {
-    return (
-      <TouchableOpacity
-        style={{
-          height: 80,
-          width: 350 / 3,
-          alignItems: "center",
-        }}
-
-        onPress={this._onPress}
-      >
-        <Text>
-          {this.short_month_texts[this.props.data.month]}
-        </Text>
-
-        <Text>
-          {`Week ${this.props.data.start_week} - ${this.props.data.end_week}`}
-        </Text>
-      </TouchableOpacity>
-    )
-  }
-}
 
 class SummaryHolder extends React.PureComponent {
 
@@ -1113,12 +186,10 @@ class SummaryHolder extends React.PureComponent {
 
     if (this.props.week_stats !== prevProps.week_stats) {
       this.updateWeekTaskCompletions(this.props.month, this.props.year, Map(this.props.week_stats))
-      console.log(this.props.week_stats)
     }
 
     if (this.props.month_stats !== prevProps.month_stats) {
       this.updateMonthTaskCompletions(this.props.month, this.props.year, Map(this.props.month_stats))
-      console.log(this.props.month_stats)
     }
 
     if (this.props.month !== prevProps.month || this.props.year !== prevProps.year) {
@@ -1207,10 +278,10 @@ class SummaryHolder extends React.PureComponent {
   }
 }
 
-class Charts extends React.PureComponent {
+class ChartSection extends React.PureComponent {
   data = [
     {
-      month: new Date(2015, 0, 1),
+      month: 1000,
       apples: 3840,
       bananas: 1920,
       cherries: 960,
@@ -1218,21 +289,21 @@ class Charts extends React.PureComponent {
       oranges: 400,
     },
     {
-      month: new Date(2015, 1, 1),
+      month: 1000,
       apples: 1600,
       bananas: 1440,
       cherries: 960,
       dates: 400,
     },
     {
-      month: new Date(2015, 2, 1),
+      month: 1000,
       apples: 640,
       bananas: 960,
       cherries: 3640,
       dates: 400,
     },
     {
-      month: new Date(2015, 3, 1),
+      month: 1000,
       apples: 3320,
       bananas: 480,
       cherries: 640,
@@ -1243,8 +314,37 @@ class Charts extends React.PureComponent {
   x_data = [0, 1, 2, 3]
   y_data = [10, 11]
 
-  colors = ['#7b4173', '#a55194', '#ce6dbd', '#de9ed6']
-  keys = ['apples', 'bananas', 'cherries', 'dates']
+  colors = ['red', '#7b4173', '#a55194', '#ce6dbd', '#de9ed6']
+  keys = ['month', 'apples', 'bananas', 'cherries', 'dates']
+
+  state = {
+    current_annotation_index: 0,
+    last_annotation_index: -1,
+
+    calendar_chosen_bool: false,
+  }
+
+  chooseAnnotation = (index) => {
+    if (this.state.current_annotation_index !== index) {
+      this.setState(prevState => ({
+        last_annotation_index: prevState.current_annotation_index,
+        current_annotation_index: index
+      }))
+    }
+  }
+
+  chooseCalendar = () => {
+    this.setState({
+      calendar_chosen_bool: true
+    })
+  }
+
+  dismissCalendar = () => {
+    this.setState({
+      calendar_chosen_bool: false
+    })
+  }
+
   render() {
     return (
       <View>
@@ -1259,6 +359,109 @@ class Charts extends React.PureComponent {
 
         <View
           style={{
+            marginTop: 20,
+            justifyContent: "center",
+            alignItems: "center",
+          }}
+        >
+          <View
+            style={{
+              flexDirection: "row",
+              width: 255,
+              height: 26,
+              borderRadius: 30,
+              borderWidth: 1,
+              borderColor: "#CCCCCC"
+            }}
+          >
+            <ChartTypeAnnotation
+              index={0}
+              annotation={"Week"}
+              current_annotation_index={this.state.current_annotation_index}
+              last_annotation_index={this.state.last_annotation_index}
+              chooseAnnotation={this.chooseAnnotation}
+            />
+
+            <ChartTypeAnnotation
+              index={1}
+              annotation={"Month"}
+              current_annotation_index={this.state.current_annotation_index}
+              last_annotation_index={this.state.last_annotation_index}
+              chooseAnnotation={this.chooseAnnotation}
+            />
+
+            <ChartTypeAnnotation
+              index={2}
+              annotation={"Year"}
+              current_annotation_index={this.state.current_annotation_index}
+              last_annotation_index={this.state.last_annotation_index}
+              chooseAnnotation={this.chooseAnnotation}
+            />
+          </View>
+
+          <TouchableOpacity
+            style={{
+              marginTop: 17,
+              justifyContent: "center",
+              alignItems: "center"
+            }}
+
+            onPress={this.chooseCalendar}
+          >
+            <Text>
+              May 5 - May 11
+            </Text>
+          </TouchableOpacity>
+
+          {this.state.calendar_chosen_bool ?
+            <Modal
+              transparent={true}
+            >
+              <View
+                style={{
+                  flex: 1,
+                  justifyContent: "center",
+                  alignItems: "center",
+                  position: "relative"
+                }}
+              >
+                <TouchableOpacity
+                  style={{
+                    flex: 1,
+                    width: Dimensions.get("window").width,
+                    backgroundColor: "black",
+                    opacity: 0.5
+                  }}
+
+                  onPress={this.dismissCalendar}
+                >
+                </TouchableOpacity>
+
+                <View
+                  style={{
+                    position: "absolute",
+                    width: 338,
+                    height: 300,
+                    backgroundColor: "white",
+                    borderRadius: 20,
+                    justifyContent: "center",
+                    alignItems: "center",
+                  }}
+                >
+                  <WeekAnnotationCalendar />
+                </View>
+              </View>
+            </Modal>
+
+            :
+
+            null
+          }
+        </View>
+
+
+        <View
+          style={{
           }}
         >
           <View
@@ -1266,7 +469,7 @@ class Charts extends React.PureComponent {
               flexDirection: "row",
             }}
           >
-            <YAxis
+            {/* <YAxis
               data={this.y_data}
               contentInset={{
                 top: 20,
@@ -1276,7 +479,7 @@ class Charts extends React.PureComponent {
               style={{
                 width: 30,
               }}
-            />
+            /> */}
             <StackedBarChart
               style={{
                 height: 200,
@@ -1293,15 +496,92 @@ class Charts extends React.PureComponent {
               }}
             />
           </View>
-          <XAxis
+          {/* <XAxis
             data={this.x_data}
             style={{
               marginLeft: 30,
             }}
             contentInset={{ left: 10, right: 10 }}
-          />
+          /> */}
         </View>
       </View>
     )
   }
 }
+
+class ChartTypeAnnotation extends React.Component {
+
+  state = {
+    container_style: styles.unchosen_container,
+    text_style: styles.unchosen_text
+  }
+
+  shouldComponentUpdate(nextProps, nextState) {
+    return this.props.index === nextProps.current_annotation_index || this.props.index === nextProps.last_annotation_index
+  }
+
+  static getDerivedStateFromProps(nextProps, prevState) {
+    if (nextProps.index === nextProps.current_annotation_index) {
+      return ({
+        container_style: styles.chosen_container,
+        text_style: styles.chosen_text
+      })
+    }
+
+    else if (nextProps.index === nextProps.last_annotation_index) {
+      return ({
+        container_style: styles.unchosen_container,
+        text_style: styles.unchosen_text
+      })
+    }
+    return null
+  }
+
+
+
+  _onPress = () => {
+    this.props.chooseAnnotation(this.props.index)
+  }
+
+  render() {
+    return (
+      <TouchableOpacity
+        style={this.state.container_style}
+        onPress={this._onPress}
+      >
+        <Text
+          style={this.state.text_style}
+        >
+          {this.props.annotation}
+        </Text>
+      </TouchableOpacity>
+    )
+  }
+}
+
+
+const styles = StyleSheet.create({
+  unchosen_container: {
+    justifyContent: "center",
+    alignItems: "center",
+    flex: 1,
+    borderRadius: 30,
+    backgroundColor: "white"
+  },
+
+  chosen_container: {
+    justifyContent: "center",
+    alignItems: "center",
+    flex: 1,
+    borderRadius: 30,
+    backgroundColor: "black"
+  },
+
+  unchosen_text: {
+    color: "gainsboro"
+  },
+
+  chosen_text: {
+    color: "white"
+  }
+})
