+++ components/shared/calendar/week-calendar/WeekCalendar.Container.js
@@ -12,7 +12,6 @@ const mapStateToProps = (state, ownProps) => {
     return {}
 }
 
-
 export default connect(
     mapStateToProps,
     null
